fix: <rect> shapes not cloning in Chrome #24
After digging through `cssText` properties applied when cloning `<rect>`, it appears that `height` and `width` are the underlying root cause to the clone not rendering properly. The simplest fix I can see at first glance is to simply check if this is a `<rect>` shape and strip these properties from the `cssText` string all together. I only apply this logic to `cssText` as this issue is specific to Chrome and Chrome seems to predominantly use the `cssText` string.