Add imaging of bare TEXT_NODE elements
Per https://github.com/1904labs/dom-to-image-more/issues/62 we throw when trying to pass a non HTMLElement as the node to be converted. This required removing the jQuery library from the test suite as it morphs the inner texts.
Add test case for issue #62 (includes adjusting the unit test framework)
Add code to handle wrapping the base text node in a SPAN
Add handling of sizing when there's no valid size
Emit reasonable SVGs with the implicit size of 300x150 if not set
Cleanup a bit to move the SVG generation into the toSvg method
Fix a control image now that we're doing the sizing correctly
Make timeouts less likely in the unit tests